Book 3: monte carlo + importance sampling
There still seems to be an issue with a glass sphere, it has a white outline that isn't present in the renders in the book. It isn't present if making another shape glass. It also reminds me of Frensel reflection, but I could't pinpoint the problem in my code. Also the way the pdf is handled in the scatterRecord struct is quite bad for perf.